update type imports
