+++ types/rangy/tests/rangy-highlighter-tests.ts
