+++ types/rangy/tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "module": "commonjs",
+        "module": "node16",
         "lib": [
             "es6",
             "dom"
@@ -20,12 +20,12 @@
       "lib/rangy-selectionsaverestore.d.ts",
       "lib/rangy-serializer.d.ts",
       "lib/rangy-textrange.d.ts",
-      "test-utils.ts",
-      "rangy-tests.ts",
-      "rangy-classapplier-tests.ts",
-      "rangy-highlighter-tests.ts",
-      "rangy-selectionsaverestore-tests.ts",
-      "rangy-serializer-tests.ts",
-      "rangy-textrange-tests.ts"
+      "tests/test-utils.ts",
+      "tests/rangy-tests.ts",
+      "tests/rangy-classapplier-tests.ts",
+      "tests/rangy-highlighter-tests.ts",
+      "tests/rangy-selectionsaverestore-tests.ts",
+      "tests/rangy-serializer-tests.ts",
+      "tests/rangy-textrange-tests.ts"
     ]
 }
